Arrange the classes into separate files.
